requirements.txt
docker~=6.0.1
python-chess